+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,14 @@
 # TBD
 
+# 0.29.1
+### Changes
+* Trying to run networking partitioning methods in Kubernetes will result in an error, rather than a panic
+* Tidying up several things inside the codebase:
+    * Kubernetes network partitioning methods now return an `error` rather than panicking
+    * `PullImage` returns an error, rather than panicking, for both Docker & Kubernetes Kurtosis backends
+    * Removed some dead code
+* Remove synchronous deletes because they're too slow
+
 # 0.29.0
 ### Features
 * Implement Kubernetes-backed file artifact expansion